Some fixes and stuff

+++ cbinarystream/binary_stream.h
@@ -7,6 +7,8 @@
 #ifndef BINARY_STREAM_H
 #define BINARY_STREAM_H
 
+#include <stdint.h>
+
 typedef struct {
 	char *buffer;
 	int size;
@@ -21,45 +23,45 @@ unsigned char get_unsigned_byte(binary_stream_t *stream);
 
 char get_byte(binary_stream_t *stream);
 
-unsigned short get_unsigned_short_le(binary_stream_t *stream);
+uint16_t get_unsigned_short_le(binary_stream_t *stream);
 
-unsigned short get_unsigned_short_be(binary_stream_t *stream);
+uint16_t get_unsigned_short_be(binary_stream_t *stream);
 
-short get_short_le(binary_stream_t *stream);
+int16_t get_short_le(binary_stream_t *stream);
 
-short get_short_be(binary_stream_t *stream);
+int16_t get_short_be(binary_stream_t *stream);
 
-unsigned long get_unsigned_triad_le(binary_stream_t *stream);
+uint32_t get_unsigned_triad_le(binary_stream_t *stream);
 
-unsigned long get_unsigned_triad_be(binary_stream_t *stream);
+uint32_t get_unsigned_triad_be(binary_stream_t *stream);
 
-long get_triad_le(binary_stream_t *stream);
+int32_t get_triad_le(binary_stream_t *stream);
 
-long get_triad_be(binary_stream_t *stream);
+int32_t get_triad_be(binary_stream_t *stream);
 
-unsigned long get_unsigned_int_le(binary_stream_t *stream);
+uint32_t get_unsigned_int_le(binary_stream_t *stream);
 
-unsigned long get_unsigned_int_be(binary_stream_t *stream);
+uint32_t get_unsigned_int_be(binary_stream_t *stream);
 
-long get_int_le(binary_stream_t *stream);
+int32_t get_int_le(binary_stream_t *stream);
 
-long get_int_be(binary_stream_t *stream);
+int32_t get_int_be(binary_stream_t *stream);
 
-unsigned long long get_unsigned_long_le(binary_stream_t *stream);
+uint64_t get_unsigned_long_le(binary_stream_t *stream);
 
-unsigned long long get_unsigned_long_be(binary_stream_t *stream);
+uint64_t get_unsigned_long_be(binary_stream_t *stream);
 
-long long get_long_le(binary_stream_t *stream);
+int64_t get_long_le(binary_stream_t *stream);
 
-long long get_long_be(binary_stream_t *stream);
+int64_t get_long_be(binary_stream_t *stream);
 
-unsigned long get_var_int(binary_stream_t *stream);
+uint32_t get_var_int(binary_stream_t *stream);
 
-long get_signed_var_int(binary_stream_t *stream);
+int32_t get_signed_var_int(binary_stream_t *stream);
 
-unsigned long long get_var_long(binary_stream_t *stream);
+uint64_t get_var_long(binary_stream_t *stream);
 
-long long get_signed_var_long(binary_stream_t *stream);
+int64_t get_signed_var_long(binary_stream_t *stream);
 
 float get_float_le(binary_stream_t *stream);
 
@@ -75,45 +77,45 @@ void put_unsigned_byte(unsigned char value, binary_stream_t *stream);
 
 void put_byte(char value, binary_stream_t *stream);
 
-void put_unsigned_short_le(unsigned short value, binary_stream_t *stream);
+void put_unsigned_short_le(uint16_t value, binary_stream_t *stream);
 
-void put_unsigned_short_be(unsigned short value, binary_stream_t *stream);
+void put_unsigned_short_be(uint16_t value, binary_stream_t *stream);
 
-void put_short_le(short value, binary_stream_t *stream);
+void put_short_le(int16_t value, binary_stream_t *stream);
 
-void put_short_be(short value, binary_stream_t *stream);
+void put_short_be(int16_t value, binary_stream_t *stream);
 
-void put_unsigned_triad_le(unsigned long value, binary_stream_t *stream);
+void put_unsigned_triad_le(uint32_t value, binary_stream_t *stream);
 
-void put_unsigned_triad_be(unsigned long value, binary_stream_t *stream);
+void put_unsigned_triad_be(uint32_t value, binary_stream_t *stream);
 
-void put_triad_le(long value, binary_stream_t *stream);
+void put_triad_le(int32_t value, binary_stream_t *stream);
 
-void put_triad_be(long value, binary_stream_t *stream);
+void put_triad_be(int32_t value, binary_stream_t *stream);
 
-void put_unsigned_int_le(unsigned long value, binary_stream_t *stream);
+void put_unsigned_int_le(uint32_t value, binary_stream_t *stream);
 
-void put_unsigned_int_be(unsigned long value, binary_stream_t *stream);
+void put_unsigned_int_be(uint32_t value, binary_stream_t *stream);
 
-void put_int_le(long value, binary_stream_t *stream);
+void put_int_le(int32_t value, binary_stream_t *stream);
 
-void put_int_be(long value, binary_stream_t *stream);
+void put_int_be(int32_t value, binary_stream_t *stream);
 
-void put_unsigned_long_le(unsigned long long value, binary_stream_t *stream);
+void put_unsigned_long_le(uint64_t value, binary_stream_t *stream);
 
-void put_unsigned_long_be(unsigned long long value, binary_stream_t *stream);
+void put_unsigned_long_be(uint64_t value, binary_stream_t *stream);
 
-void put_long_le(long long value, binary_stream_t *stream);
+void put_long_le(int64_t value, binary_stream_t *stream);
 
-void put_long_be(long long value, binary_stream_t *stream);
+void put_long_be(int64_t value, binary_stream_t *stream);
 
-void put_var_int(unsigned long value, binary_stream_t *stream);
+void put_var_int(uint32_t value, binary_stream_t *stream);
 
-void put_signed_var_int(long value, binary_stream_t *stream);
+void put_signed_var_int(int32_t value, binary_stream_t *stream);
 
-void put_var_long(unsigned long long value, binary_stream_t *stream);
+void put_var_long(uint64_t value, binary_stream_t *stream);
 
-void put_signed_var_long(long long value, binary_stream_t *stream);
+void put_signed_var_long(int64_t value, binary_stream_t *stream);
 
 void put_float_le(float value, binary_stream_t *stream);
