+++ podrum/network/raknet/rakpacket.c
@@ -13,9 +13,9 @@
 packet_unconnected_ping_t get_packet_unconnected_ping(binary_stream_t *stream)
 {
 	packet_unconnected_ping_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.timestamp = get_unsigned_long_be(stream);
-	stream->offset += 16; // MAGIC
+	stream->offset += 16; /* MAGIC */
 	packet.guid = get_unsigned_long_be(stream);
 	return packet;
 }
@@ -23,11 +23,11 @@ packet_unconnected_ping_t get_packet_unconnected_ping(binary_stream_t *stream)
 packet_unconnected_pong_t get_packet_unconnected_pong(binary_stream_t *stream)
 {
 	packet_unconnected_pong_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.timestamp = get_unsigned_long_be(stream);
 	packet.guid = get_unsigned_long_be(stream);
-	stream->offset += 16; // MAGIC
-	unsigned short length = get_unsigned_short_be(stream);
+	stream->offset += 16; /* MAGIC */
+	uint16_t length = get_unsigned_short_be(stream);
 	packet.message = (char *) malloc(length + 1);
 	int i;
 	for (i = 0; i < length; ++i) {
@@ -40,9 +40,9 @@ packet_unconnected_pong_t get_packet_unconnected_pong(binary_stream_t *stream)
 packet_incompatible_protocol_version_t get_packet_incompatible_protocol_version(binary_stream_t *stream)
 {
 	packet_incompatible_protocol_version_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.protocol_version = get_unsigned_byte(stream);
-	stream->offset += 16; // MAGIC
+	stream->offset += 16; /* MAGIC */
 	packet.guid = get_unsigned_long_be(stream);
 	return packet;
 }
@@ -50,17 +50,17 @@ packet_incompatible_protocol_version_t get_packet_incompatible_protocol_version(
 packet_open_connection_request_1_t get_packet_open_connection_request_1(binary_stream_t *stream)
 {
 	packet_open_connection_request_1_t packet;
-	stream->offset += 17; // PACKET_ID + MAGIC
+	stream->offset += 17; /* PACKET_ID + MAGIC */ 
 	packet.protocol_version = get_unsigned_byte(stream);
 	stream->offset = stream->size - 1;
-	packet.mtu_size = stream->size; // PACKET_ID + MAGIC + PROTOCOL_VERSION + PAD_BYTES
+	packet.mtu_size = stream->size; /* PACKET_ID + MAGIC + PROTOCOL_VERSION + PAD_BYTES */
 	return packet;
 }
 
 packet_open_connection_reply_1_t get_packet_open_connection_reply_1(binary_stream_t *stream)
 {
 	packet_open_connection_reply_1_t packet;
-	stream->offset += 17; // PACKET_ID + MAGIC
+	stream->offset += 17; /* PACKET_ID + MAGIC */ 
 	packet.guid = get_unsigned_long_be(stream);
 	packet.use_security = get_unsigned_byte(stream);
 	packet.mtu_size = get_unsigned_short_be(stream);
@@ -70,7 +70,7 @@ packet_open_connection_reply_1_t get_packet_open_connection_reply_1(binary_strea
 packet_open_connection_request_2_t get_packet_open_connection_request_2(binary_stream_t *stream)
 {
 	packet_open_connection_request_2_t packet;
-	stream->offset += 17; // PACKET_ID + MAGIC
+	stream->offset += 17; /* PACKET_ID + MAGIC */ 
 	packet.address = get_misc_address(stream);
 	packet.mtu_size = get_unsigned_short_be(stream);
 	packet.guid = get_unsigned_long_be(stream);
@@ -80,7 +80,7 @@ packet_open_connection_request_2_t get_packet_open_connection_request_2(binary_s
 packet_open_connection_reply_2_t get_packet_open_connection_reply_2(binary_stream_t *stream)
 {
 	packet_open_connection_reply_2_t packet;
-	stream->offset += 17; // PACKET_ID + MAGIC
+	stream->offset += 17; /* PACKET_ID + MAGIC */ 
 	packet.guid = get_unsigned_long_be(stream);
 	packet.address = get_misc_address(stream);
 	packet.mtu_size = get_unsigned_short_be(stream);
@@ -91,10 +91,10 @@ packet_open_connection_reply_2_t get_packet_open_connection_reply_2(binary_strea
 packet_acknowledge_t get_packet_acknowledge(binary_stream_t *stream)
 {
 	packet_acknowledge_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.sequence_numbers_count = 0;
-	packet.sequence_numbers = (unsigned long *) malloc(0);
-	unsigned short record_count = get_unsigned_short_be(stream);
+	packet.sequence_numbers = (uint32_t *) malloc(0);
+	uint16_t record_count = get_unsigned_short_be(stream);
 	int i;
 	unsigned char is_single;
 	int index;
@@ -103,12 +103,12 @@ packet_acknowledge_t get_packet_acknowledge(binary_stream_t *stream)
 		is_single = get_unsigned_byte(stream);
 		if (is_single != 0) {
 			++packet.sequence_numbers_count;
-			packet.sequence_numbers = (unsigned long *) realloc(packet.sequence_numbers, packet.sequence_numbers_count * sizeof(long));
+			packet.sequence_numbers = (uint32_t *) realloc(packet.sequence_numbers, packet.sequence_numbers_count * sizeof(uint32_t));
 		 	packet.sequence_numbers[packet.sequence_numbers_count - 1] = get_unsigned_triad_le(stream);
 		} else {
 			index = get_unsigned_triad_le(stream);
 			end_index = get_unsigned_triad_le(stream);
-			packet.sequence_numbers = (unsigned long *) realloc(packet.sequence_numbers, (packet.sequence_numbers_count + (end_index - index + 1)) * sizeof(long));
+			packet.sequence_numbers = (uint32_t *) realloc(packet.sequence_numbers, (packet.sequence_numbers_count + (end_index - index + 1)) * sizeof(uint32_t));
 			while (index <= end_index) {
 				packet.sequence_numbers[packet.sequence_numbers_count] = index;
 				++packet.sequence_numbers_count;
@@ -122,7 +122,7 @@ packet_acknowledge_t get_packet_acknowledge(binary_stream_t *stream)
 packet_frame_set_t get_packet_frame_set(binary_stream_t *stream)
 {
 	packet_frame_set_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.sequence_number = get_unsigned_triad_le(stream);
 	packet.frames_count = 0;
 	packet.frames = (misc_frame_t *) malloc(0);
@@ -138,7 +138,7 @@ packet_frame_set_t get_packet_frame_set(binary_stream_t *stream)
 packet_connection_request_t get_packet_connection_request(binary_stream_t *stream)
 {
 	packet_connection_request_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.guid = get_unsigned_long_be(stream);
 	packet.timestamp = get_unsigned_long_be(stream);
 	return packet;
@@ -147,7 +147,7 @@ packet_connection_request_t get_packet_connection_request(binary_stream_t *strea
 packet_connection_request_accepted_t get_packet_connection_request_accepted(binary_stream_t *stream)
 {
 	packet_connection_request_accepted_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.address = get_misc_address(stream);
 	packet.system_index = get_unsigned_short_be(stream);
 	int i;
@@ -162,7 +162,7 @@ packet_connection_request_accepted_t get_packet_connection_request_accepted(bina
 packet_new_incoming_connection_t get_packet_new_incoming_connection(binary_stream_t *stream)
 {
 	packet_new_incoming_connection_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.address = get_misc_address(stream);
 	int i;
 	for (i = 0; i < 20; ++i) {
@@ -176,7 +176,7 @@ packet_new_incoming_connection_t get_packet_new_incoming_connection(binary_strea
 packet_connected_ping_t get_packet_connected_ping(binary_stream_t *stream)
 {
 	packet_connected_ping_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.timestamp = get_unsigned_long_be(stream);
 	return packet;
 }
@@ -184,7 +184,7 @@ packet_connected_ping_t get_packet_connected_ping(binary_stream_t *stream)
 packet_connected_pong_t get_packet_connected_pong(binary_stream_t *stream)
 {
 	packet_connected_pong_t packet;
-	++stream->offset; // PACKET_ID
+	++stream->offset; /* PACKET_ID */
 	packet.request_timestamp = get_unsigned_long_be(stream);
 	packet.reply_timestamp = get_unsigned_long_be(stream);
 	return packet;
@@ -258,7 +258,7 @@ void put_packet_open_connection_reply_2(packet_open_connection_reply_2_t packet,
 
 void put_packet_acknowledge(packet_acknowledge_t packet, int opts, binary_stream_t *stream)
 {
-	// Sort out the sequence numbers
+	/* Sort out the sequence numbers */
 	int temp = 0;
 	int i;
 	for (i = 0; i < packet.sequence_numbers_count; ++i) {
@@ -278,11 +278,11 @@ void put_packet_acknowledge(packet_acknowledge_t packet, int opts, binary_stream
 	temp_stream.buffer = (char *) malloc(0);
 	int record_count = 0;
 	if (packet.sequence_numbers_count > 0) {
-		long start_index = packet.sequence_numbers[0];
-		long end_index = packet.sequence_numbers[0];
+		uint32_t start_index = packet.sequence_numbers[0];
+		uint32_t end_index = packet.sequence_numbers[0];
 		for (i = 1; i < packet.sequence_numbers_count; ++i) {
-			long current_index = packet.sequence_numbers[i];
-			long diff = current_index - end_index;
+			uint32_t current_index = packet.sequence_numbers[i];
+			uint32_t diff = current_index - end_index;
 			if (diff == 1) {
 				end_index = current_index;
 			} else if (diff > 1) {
